// newflow-main/packages/frontend/editor-ui/src/Interface.ts
export interface HistoryState {
	redoStack: Undoable[];
	undoStack: Undoable[];
	currentBulkAction: BulkCommand | null;
	bulkInProgress: boolean;
}